Refactor PSI structure.
