+++ assets/js/modules/pagespeed-insights/components/dashboard/DashboardPageSpeedCTA.js
@@ -29,12 +29,12 @@ import {
 	activateOrDeactivateModule,
 	getReAuthURL,
 	showErrorNotification,
-} from '../../../util';
-import CTA from '../../../components/notifications/cta';
-import data from '../../../components/data';
-import GenericError from '../../../components/notifications/generic-error';
-import { STORE_NAME as MODULES_STORE } from '../../../googlesitekit/modules/datastore/constants';
-import { STORE_NAME as USER_STORE, PERMISSION_MANAGE_OPTIONS } from '../../../googlesitekit/datastore/user/constants';
+} from '../../../../util';
+import CTA from '../../../../components/notifications/cta';
+import data from '../../../../components/data';
+import GenericError from '../../../../components/notifications/generic-error';
+import { STORE_NAME as MODULES_STORE } from '../../../../googlesitekit/modules/datastore/constants';
+import { STORE_NAME as USER_STORE, PERMISSION_MANAGE_OPTIONS } from '../../../../googlesitekit/datastore/user/constants';
 const { useSelect } = Data;
 
 function DashboardPageSpeedCTA() {
