+++ assets/js/modules/pagespeed-insights/components/dashboard/DashboardPageSpeedWidget.js
@@ -26,7 +26,7 @@ import { string } from 'prop-types';
  */
 import Data from 'googlesitekit-data';
 import Widgets from 'googlesitekit-widgets';
-import { STORE_NAME as MODULES_STORE } from '../../../googlesitekit/modules/datastore/constants';
+import { STORE_NAME as MODULES_STORE } from '../../../../googlesitekit/modules/datastore/constants';
 import DashboardPageSpeedCTA from './DashboardPageSpeedCTA';
 import DashboardPageSpeed from './DashboardPageSpeed';
 const { useSelect } = Data;
