+++ assets/js/modules/pagespeed-insights/util/index.js
@@ -19,7 +19,7 @@
 /**
  * WordPress dependencies
  */
-import { CATEGORY_FAST, CATEGORY_AVERAGE, CATEGORY_SLOW } from '../constants';
+import { CATEGORY_FAST, CATEGORY_AVERAGE, CATEGORY_SLOW } from './constants';
 
 /**
  * Retrieve the score category slug based on the given score.
